support match(json) for non-store requests
